apps/api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pyotp>=2.9.0
qrcode[pil]>=7.4.0
python-multipart>=0.0.9
httpx>=0.27.0
aiofiles>=23.2.0
cryptography>=42.0.0
redis>=5.0.0
ruff>=0.3.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
anyio>=4.0.0
pillow>=10.0.0
